Added folder Footers and moved Footer files
