+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import Footer from "./components/Footer.js";
+import Footer from "./components/Footers/Footer.js";
 
 function App() {
     return (
